Software/DroneSoftware/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pymavlink==2.4.14
Adafruit-Blinka==8.20.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.3
adafruit-circuitpython-requests==2.0.0
adafruit-circuitpython-ssd1306==2.12.14
adafruit-circuitpython-typing==1.9.4
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.47.0
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
blinker==1.4
Jetson.GPIO==2.1.1
MAVProxy==1.8.62
paho-mqtt==1.6.1
Pillow==7.0.0
pyserial==3.5
